PLFM-7194: Edited integration test after method

+++ services/repository-managers/src/test/java/org/sagebionetworks/repo/manager/file/MultipartManagerV2ImplAutowireTest.java
@@ -204,7 +204,9 @@ public class MultipartManagerV2ImplAutowireTest {
 			if (fileHandle instanceof S3FileHandle) {
 				s3Client.deleteObject(fileHandle.getBucketName(), fileHandle.getKey());
 			} else if (fileHandle instanceof GoogleCloudFileHandle) {
-				gcClient.deleteObject(fileHandle.getBucketName(), fileHandle.getKey());
+				if (gcClient.doesObjectExist(fileHandle.getBucketName(), fileHandle.getKey())) {
+					gcClient.deleteObject(fileHandle.getBucketName(), fileHandle.getKey());
+				}
 			}	
 		}
